requirements.txt
Pillow==8.2.0
pyttsx3==2.90
SpeechRecognition==3.8.1
numpy==1.21.2
scikit-learn==0.24.2
